Organized configurations. Added XClockCompositeDiagram.
