+++ examples/demo/clock/src/clock/x/XClockCompositeDiagram.java
@@ -1,0 +1,17 @@
+package clock.x;
+
+import hu.elte.txtuml.api.layout.CompositeDiagram;
+import hu.elte.txtuml.api.layout.East;
+import hu.elte.txtuml.api.layout.North;
+import hu.elte.txtuml.api.layout.Spacing;
+import clock.x.model.HourHand;
+import clock.x.model.MinuteHand;
+import clock.x.model.SecondHand;
+import clock.x.model.Clock;
+
+public class XClockCompositeDiagram extends CompositeDiagram<Clock> {
+	@East(from=HourHand.hourHand.class, val=MinuteHand.minuteHand.class)
+	@North(from=MinuteHand.minuteHand.class, val=SecondHand.secondHand.class)
+	@Spacing(0.4)
+	class ClockLayout extends Layout {}
+}
