--- examples/demo/machine/src/machine1/j/DefaultConfiguration.java
@@ -1,10 +1,0 @@
-package machine1.j;
-
-import hu.elte.txtuml.api.deployment.Configuration;
-import hu.elte.txtuml.api.deployment.Group;
-import machine1.j.model.Machine;
-import machine1.j.model.User;
-
-@Group(contains = { Machine.class, User.class })
-public class DefaultConfiguration extends Configuration {
-}
